pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["sena1267 <silc0504@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.7"
python-jose = "^3.3.0"
uvicorn = "^0.21.1"
passlib = "^1.7.4"
python-dotenv = "^1.0.0"
pydantic = "^1.10.6"
fastapi = "^0.95.0"
python-multipart = "^0.0.6"
bcrypt = "^4.0.1"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
flake8 = "^6.0.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
